+++ src/autocomplete-options.ts
@@ -10,6 +10,16 @@ export default class AutocompleteOptions {
     name: string;
 
     /**
+     * Specify source options source
+     * @example ['Afghanistan', 'Albania', 'Algeria', ...more]
+     * @example [{ label: 'Afghanistan', value: 'AFG' }, ...more]
+     * @example 'https://some-endpoint.somewhere/available'
+     * @example (query, render, isFirstCall) => render(arrayToUse)
+     * @example (query) => async () => arrayToUse
+     */
+    source: string[] | any[] | string | Function | Promise<any[]>;
+
+    /**
      * Properties to use for label and value when source is an Array of Objects
      */
     sourceMapping: any = {};
@@ -19,6 +29,12 @@ export default class AutocompleteOptions {
      * `label` will always be used
      */
     alsoSearchIn: string[] = [];
+
+    /**
+     * If no exact match is found,
+     * create an entry in the options list for the current search text
+     */
+    create: boolean | ((value: string) => string | object) = false;
 
     /**
      * Input delay after typing before running a search
@@ -131,9 +147,9 @@ export default class AutocompleteOptions {
 
     /**
      * Automatically clear the screen reader announcement element after the specified delay
-     * Defaults to 2 seconds if true
+     * Number is in milliseconds. If true, defaults to 5000.
      */
-    srAutoClear: boolean | number = false;
+    srAutoClear: boolean | number = 5000;
 
     /**
      * Screen reader text used in multiple mode for element deletion.