+++ webpack.config.js
@@ -6,16 +6,16 @@ module.exports = {
             {
                 test: /\.tsx?$/,
                 use: 'ts-loader',
-                exclude: /node_modules/
-            }
-        ]
+                exclude: /node_modules/,
+            },
+        ],
     },
     resolve: {
-        extensions: ['.tsx', '.ts', '.js']
+        extensions: ['.tsx', '.ts', '.js'],
     },
     output: {
         libraryTarget: 'umd',
         path: __dirname + '/dist',
-        filename: 'aria-autocomplete.min.js'
-    }
+        filename: 'aria-autocomplete.min.js',
+    },
 };